# perf: load image to array in cache sizes