This PR fills in images in 'mini-batches' (actually mini-chunks of each image) to avoid going over  the standard L2-cache. This provides almost 2x performance and less memory allocations.

Benchmark
```
from jina.types.document.converters import to_image_blob
%timeit im_numpy = to_image_blob(`path_to_photo.jpg`)
```

Master branch
```
203 ms ± 12 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

This PR:
```
114 ms ± 1.79 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
```