@JoanFM 2021-09-26
As discussed, this optimizations should be moved maybe in some specific `Executor` and not being taken care by `core`